Split classes into seperate file, smarter AI that knows how not to lose(but doesn't know how to win)
