+++ connect_four/connect_four.rb
@@ -1,33 +1,6 @@
-class Player
+require './player.rb'
+require './board.rb'
 
-  def initialize(is_ai = false)
-    @is_ai = is_ai
-  end
-
-  def get_move
-    @is_ai? ai_move : human_move
-  end
-
-  def human_move
-    puts "Input a move Col 1 to 6"
-    col_to_add_to = gets.chomp.to_i
-    col_to_add_to -= 1
-    until col_to_add_to >= 0 && col_to_add_to < 6
-      puts "Invalid column number, try again!"
-    	col_to_add_to = self.human_move
-    end
-    col_to_add_to
-  end
-
-  # Algorithm for deciding the ai's move
-  def ai_move
-  	(0..5).to_a.sample
-  end
-
-end
-
-# see if there is a winning move
-# stop human if they are winning
 
 
 class Game
@@ -56,12 +29,12 @@ class Game
 
   def get_valid_move
     puts "Turn #{@turn + 1}, Player #{(@turn % 2) + 1}"
-  	col = current_player.get_move
+  	col = current_player.move
   	if @board.col_full?(col)
       puts "That column is full."
       @board.display
       while @board.col_full?(col)
-  		  col = current_player.get_move
+  		  col = current_player.move
         puts "That column is full."
         @board.display
       end
@@ -76,103 +49,20 @@ class Game
       puts "Do you want to go first or second?"
       choice = gets.chomp
       if choice == "1"
-        @players << Player.new
-        @players << Player.new(true)
+        @players = [Human.new, AI.new(@board, 1)]
       else
-        @players << Player.new(true)
-        @players << Player.new
+        @players = [AI.new(@board, 0), Human.new]
       end
     else
-      @players << Player.new
-      @players << Player.new
+      @players = [Human.new, Human.new]
     end
   end
 
-end
-
-class Board
-
-  def initialize
-  	@state = [[],[],[],[],[],[]]
-  end
-
-  def move(args)
-  	@state[args[0]] << args[1]
-  end
-
-  # Returns the winner, or NIL if there is no winner
-  def winner
-    # Horizontal
-    0.upto(5) do |row|
-      0.upto(2) do |col|
-        if (@state[0 + col][row] == @state[1 + col][row]) &&
-        (@state[0 + col][row] == @state[2 + col][row]) &&
-        (@state[0 + col][row] == @state[3 + col][row])
-          return @state[col][row] unless @state[col][row].nil?
-        end
-      end
-    end
-
-    # Vertical
-    0.upto(5) do |col|
-      0.upto(2) do |row|
-        if (@state[col][0 + row] == @state[col][1 + row]) &&
-        	(@state[col][0 + row] == @state[col][2 + row]) &&
-        	(@state[col][0 + row] == @state[col][3 + row])
-          return @state[col][row] unless @state[col][row].nil?
-        end
-      end
-    end
-
-    # Bottom left to upper right
-    0.upto(2) do |row|
-      0.upto(2) do |col|
-        if (@state[col][row] == @state[1 + col][1 + row]) &&
-        	(@state[col][row] == @state[2 + col][2 + row]) &&
-        	(@state[col][row] == @state[3 + col][3 + row])
-          return @state[col][row] unless @state[col][row].nil?
-        end
-      end
-    end
-
-    # Upper left to lower right
-    5.downto(3) do |row|
-      0.upto(2) do |col|
-        if (@state[col][row] == @state[col + 1][row - 1]) &&
-        	(@state[col][row] == @state[col + 2][row - 2]) &&
-        	(@state[col][row] == @state[col + 3][row - 3])
-          return @state[col][row] unless @state[col][row].nil?
-        end
-      end
-    end
-
-    return nil
-  end
-
-  def full?
-    return @state.reduce(0){|acc, col| acc += col.length} == 36
-  end
-
-  def display
-    puts ""
-    puts ""
-    5.downto(0) do |idx|
-      0.upto(5) do |pos|
-        @state[pos][idx]? (print "#{@state[pos][idx]}") : (print " ")
-      end
-      puts ""
-    end
-  end
-
-  def col_full?(col_no)
-  	@state[col_no].length > 5
-  end
 end
 
 game = Game.new
 
 game.play
-
 
 =begin
